Refactor App component and add Dashboard with summary features
