+++ src/App.js
@@ -1,20 +1,14 @@
 // import logo from './logo.svg';
 import './App.css';
 // import Dashboard from './Dashbord/Dashboard';
-
-import Dashboard from './Components/Dashboard/Dashboard';
-// import SignIn from './Components/SignIn';
-// import SignUp from './Components/SignUp';
-
-
+import Dashboard from './Components/Dashbord/Dashboard'
+import SignIn from './Components/Authentication/SignIn';
+import SignUp from './Components/Authentication/SignUp';
 function App() {
   return (
     <div className="App">
-
-      <Dashboard/>
-
-      {/* <SignIn/>
-      <SignUp /> */}
+      <SignIn/>
+      <SignUp />
       <Dashboard />
 
     </div>
